+++ src/integrations/lark/lark.client.ts
@@ -35,6 +35,7 @@ export class LarkClient {
   public readonly wiki: lark.Client['wiki'];
   public readonly vc: lark.Client['vc'];
   public readonly minutes: lark.Client['minutes'];
+  public readonly wsClient: lark.WSClient;
 
   constructor(
     @Inject(larkConfig.KEY) private readonly cfg: ConfigType<typeof larkConfig>,
@@ -82,6 +83,7 @@ export class LarkClient {
     this.wiki = this.client.wiki;
     this.vc = this.client.vc;
     this.minutes = this.client.minutes;
+    this.wsClient = this.wsClient;
 
     this.logger.log('Lark client initialized successfully');
   }
